Retire any reference to a specific IP address
And clarify security policy of the framework.
In particulary we can have 2 cases:

1) User app (like a worker task) runs locally on the same
   machine of the other components (Celery, MangoDb, etc)

2) User app runs on another machine and acceses the host
   through a network connection.

In the first case there is nothing to do, in the second case
user shall setup a ssh connection with port forwarding, see
README for details.

In both cases fishtest is agonstic about the remote host and
even does not know (or care) if the components are remote or
local.