Remove start_worker.sh from repo
This file is not Windows compatible and anyhow is
an extra, not needed, intermediate step. User very
probably will already wrap the call to worker.py
in a script (.sh or .bat) togheter with access credentials
that of course cannot be stored as part of git repo.

README.md already does not contain any reference
to this file and asks the user to directly run
worker.py from its directory.