Fix run_games exception
Currently after the first exception we send api/failed_task
to host that set the task as not active. So when we retry (up
to five times) and in case succed, when we send updated_task
result is ignored because task is not active anymore.

So fix it allowing the worker to retry forever but redo the
full loop, starting from getting a new task.