Change default book to 8moves_GM.pgn
As Adam Hair says:

"I can share with the group a pgn of 48,491 games that are 8 moves deep.
These games came from GM games that Norm Pollock collected and shares at
his website (the same games that Martin used to make the TCEC book used
in stage 1). I used some of the top engines to filter out positions that
may be too unbalanced. The final position of each truncated game is unique,
so that it is a source of 48,491 unique starting positions. Also, I used
some of the top engines to filter out positions that may be too unbalanced.
It is not perfect, but it may be suitable for the group's purposes.