First cut at adding CLOP interface
Added a new glue script that support multiple
functions according to the command line arguments:

1) start_clop: Starts CLOP with given parameters
2) write_result: Updates the result of a played game
3) Called by CLOP to start a new game

There is a buffer (a clopdb) that disjoints the script
from the fishtest framework. In particular games to run
are queued in clopdb, in async fashion fishtest peeks from
clopdb a game to run, runs it and then updates the result
in the same clopdb row. Once result is updated is feeded
back to CLOP and corresponding row is removed from the DB