Increase the time waiting for a new game result. This should increase the maximum allowed game duration to 30 minutes, and allows tests with longer time controls. Also bump up worker version.
This change worked recently for CoffeeOne.
I propose to run a test of about 1,000 games at 5min+3sec.

I chose a limit of 30 minutes as I think this is a good compromise
between the maximum game duration, and the time an unattended PC
must wait until the hanging process is being killed.