Feed a random seed to cutechess.
Cutechess presently uses the current time, in seconds, to seed its RNG.
This makes it very easy for various machines in fishtest to choose the
same random seed on the same workload, particularly when a large number
of machines join a machine pool for a specific test.

The fix here is to pass a random integer to cutechess via the "-srand"
option.  A platform-independent seed is chosen using os.urandom(),
which raises an exception if a randomness source is unavailable.

Cutechess only uses 32 bits of random state (a "quint32"), and the
parser for -srand requires an unsigned integer (although the
Mersenne::initialize routine just takes an int), so the routine here
generates an unsigned integer.

Issue originally discovered by VoyagerOne at
https://groups.google.com/forum/#!topic/fishcooking/CW7Vlwur8CQ