treat boundaries symmetrically in spsa
prior to this patch, spsa boudnary treatment has the following two
properties:
* samples on either side of the iterate are clipped independently,
  meaning the sampling distribution is not symmetric about the current
  parameter.
* After clipping, it's possible to have a value exactly at the boundary.

This is bad because spsa interprets a draw as "do nothing", meaning
parameters are not updated.  but that only makes sense if the present
parameters are the average of the parameters tried in the test, which is
not the clase with the old clipping.

The second point is bad if one believes the first point is bad (causes
excessive asymmetry).

So this patch introduces clipping which is always symmetric.  It is
moreover cautious to not move all the way to the boundary when doing
this update.