randomized rounding for spsa
spsa converts floating point parameter values to integers via round(),
which is dicontinuous.  This patch adds randomized rounding,
which converts x+p (where x is integer and p is within [0,1)) to
x with probability 1-p and x+1 with probability p (thus we get back x+p
in expectation), which is continuous.