Increase default SPRT game limit to 1000000
Currently, SPRT tests have a 128000 game limit, meaning that at most 128
workers can be assigned to a test, unless the limit is manually
increased. With the recent increase of fishtest machines, we are hitting
this limit quite often. The 128 chunks can get assigned quickly when the
test has barely started. As a result, some machines become idle.

This patch allows 1000 chunks to be assigned to a test by default.