Cache Run data and rewrite request_task in Python
The idea is simple, there is no need to read Run data more than once
from MongoDb, because we are the only one updating.

The only exceptions are:

* request_task, which executes a MongoDb query that modifies the database,
  so the Run data cache is invalidated.
* scavenge.py (new internal implementation for active Runs in this PR)
* update_users.py (does not write Run data)

This PR will also reduce the write frequency from update_task which is
executed hundreds of times each second and overwrites the data it has
just written a few milliseconds earlier. These writes are now buffered
and a buffer is flushed every second.

In addition rewrite and speedup request_task by executing the selection
logic in Python.