Retire worker reconnect logic
When a worker disconnects from fishtest because of errors
and reconnects without being restarted (it gets a new ID
on restart) it is handed it's old job, unless it was
already assigned to another worker.

This situation is rare and the logic requires scanning all
tasks handed out.

https://github.com/glinscott/fishtest/pull/186 increased
the speed by remembering active task for each worker,
but this information is lost on a fishtest reboot.

Benefit from the current code is minimal so retire it.