Add exception handling of download & build process
In case of an exception occurred during the download & build process, add a catch block so that temp directory cleanup is not skipped.

It is also possible to retry a few times here instead of going through the expensive request_task process in case of a recoverable failure in the future improvements.