Catch the kill process error
Print an error message and pass the execution if the kill of the process fails.
The kill of the process is used as safety measure against a hanging process.
At the moment all the running processes seems to run and close fine,
so we are always trying to kill an already closed process.
With a WSL worker that run a SPSA test when the kill fails is raised
an exception that stops the execution for some seconds slowing the test.

Refactoring of 'try: except:' code as suggested by @tomtor
Change of the message as suggested by @vondele and @Stefano80