Keep the zipped MongoDB dump
The size of the DB is increasead, so keep locally a zipped copy of
the MongoDB dump in order to have more free HD space for the server maintenance.
This speed up also the copy of the DB on the development server.

Load the AWS variables

The file .bashrc is loaded automatically only with an interactive shell.
With a script scheduled by cron we need to source the .bashrc .
Bug introduce with https://github.com/glinscott/fishtest/pull/285