Make the worker calculate the pentanomial frequencies.
There are no server changes in this patch.

The updated worker will send the pentanomial frequencies
to the server and the server will faithfully store them in
its databaese as it does with all information received from
the client.

The actual logic added to the worker is quite minor.
A dictionary "rounds" keeps track of the finished games. For every
finished game pair (which consists of an odd numbered game
and the next even numbered game) the pentanomial frequencies
are updated.

Since sending incorrect frequencies to the server would
be quite distastrous, a lot of sanity checking is done to
make sure that the pentanomial frequencies are consistent
with the trinomial frequencies provided by cutechess itself.