Make the worker calculate the pentanomial frequencies.
There are no server changes in this patch.

The updated worker will send the pentanomial frequencies
to the server and the server will faithfully store them in
its databaese as it does with all information received from
the client.

I checked that the server changes necessary to make
effective use of the pentanomial frequencies are quite minor.
This will be the subject of some subsequent PR's.

I am making this PR now so that the worker changes can be reviewed.

The actual logic added to the worker is quite minor.
A dictionary "rounds" keeps track of the finished games. For every
finished game pair (which consists of an odd numbered game
and the next even numbered game) the pentanomial frequencies
are updated.

Since sending incorrect frequencies to the server would
be quite distastrous, a lot of sanity checking is done to
make sure that the pentanomial frequencies are consistent
with the trinomial frequencies provided by cutechess itself.

I have tested the updated worker on my local copy of fishtest.