Include the pentanomial frequencies in the "elo results" widget.
The pentanomial frequencies cannot be used solely behind the scenes
since that would make it difficult to verify independently the
statistical quantities derived from them.

So here we include them in the elo widget. As yet no statistical
quantities are computed.

My next plan is to apply the pentanomial frequencies to fixed length
tests. That would be the first statistical application.