Handle test stop, restart in statistics update
When a test is stopped and restarted (by a modify) this
can cause many runs (30 days period) to be counted twice.
This happens only when the restarted test was the only new
finished test at the last execution of this script.
When the test is restarted before the script is executed
then there is also no error condition.

This can be fixed by storing also all older run-ids,
with a small performance impact (bigger MongoDb table).

Note that tasks in the restarted test are counted twice
when the error condition happens. This should be very rare.