Small SPSA speedup with several code improvements
Clip a value in a range using an if statement instead that a double function calls.
Drop some if statements and rewrote the rounding code as an oneliner.
The new rounding code has the same probabilistic behavior of the old code, results are equal if mirroring the random
variable (but the code will be a little slower).
Code simplification removing unnecessary variables.