Various internal/external cleanups for the SPRT calculator.
There are now minor and major vertical grid lines.

The Elo increment used for graph drawing is now
a bit adaptive. This gives smoother tooltip behaviour
for small elo differences.

All mouse events are now captured by a div specially created
for that purpose. The GC default event handlers are
thus completely disabled. This removes some prior artifacts, e.g.
when the user clicked on the graphs.

Events are now disabled until the charts are fully drawn.

In general event handling is greatly simplified internally.

Some html cleanup. In particular the name attribute
of the input fields is not used and thus was removed.
Most inline styling was removed. A spurious </div> was
removed.