Fix the worker update process
Update only the worker files so that the update process will work
also with a custom worker folder name.
Enforce the download of the lastest version of cutechess-cli, books etc.
renaming the "testing" folder with a timestamp. Keep the lastest 3 backups
of the "testing" folder to save eventual custom user files.
Thanks to @vondele for some ideas and code.