Better overshoot correction.
The theoretical properties of an SPRT are established under the
assumption that updates are continuous.

To have accurate agreement with the predictions, when the updates are
discrete, one has to estimate the expected "overshoot" and terminate
the SPRT a bit early, accordingly.  Unfortunately, to the best of our
knowledge, overshoot estimation is (surprisingly) computationally expensive.

Moreover in the case of a GSPRT there is the added complication of
unknown parameters and these have to be continuously estimated from
the sample. This more or less means that the expected overshoot
estimation has to be performed after every update, which is
prohibitively expensive (IMHO, doing it only when we are "close" to
one of the boundaries would lead to messy and possibly fragile code).

A theoretical formula for the expected overshoot is: Siegmund -
Sequential Analysis - Corollary 8.33. Like other such formulas, it is
difficult to evaluate theoretically, but the result can be very easily
estimated, with only a few lines of code, from the ongoing test itself.
This "dynamic overshoot estimation" is implemented in this commit.

That this works very well can be verified by simulation. A multi-threaded
simulator written in C can be found here.

https://github.com/vdbergh/simul

Example:

$ ./simul --elo0 0 --elo1 10 --elo 0

yields

...
sims=20001904 pass=0.050322[0.050175,0.050469] ...
...

whereas

$ ./simul --elo0 0 --elo1 10 --elo 0 --noovcor

yields

...
sims=20000943 pass=0.048530[0.048385,0.048674] ...
...

(the confidence interval is 3 sigma).

Obviously the version with overshoot correction is much more accurate.

Currently Fishtest uses an approximate formula for overshoot
correction which is actually not even asymptotically correct, although
it works well for small Elo differences. However for larger Elo
differences it starts to break down. This is undesirable as people may
copy the Fishtest code and apply it in a different context.

In the implementation the SPRT() function in stat_util.py now takes an
sprt dictionary as argument. This dictionary may optionally contain a
subdictionary with dynamic overshoot data which is updated in-place.
It is assumed that SPRT() is called after each finished game (trinomial)
or game pair (pentanomial). But calling SPRT() multiple times with
the same parameters is safe and skipped updates are also handled sensibly.
After a purge the overshoot data is deleted. Deleting the overshoot data
is always safe as it simply means that the test is carried out with error
probabilities that are slightly smaller than the assigned ones.

The actual overshoot estimates can be found on the raw statistics
page. Note that the estimate corresponding to the test outcome (H0 or H1)
is typically the most accurate.