Worker changes to allow batching.
A new quantity "batch_size" is defined.

In principle batch_size is the number of games that have to be
completed by the worker before an update is sent to the server.
However for sprt and numgames tests it is actually twice the
number of completed game pairs.

batch_size for sprt is communicated by the server. The reason
is that for the sprt algorithm to be theoretically correct all
workers have to use the same batch size.

For other tests batch_size is currently set by the worker to twice the
number of games it is capable of running concurrently.

This is equal to the number of cores reserved for the worker,
divided by the number of threads used by the engines.