Server changes to set the batch size for sprt tests.
In this PR the batch size is set to 8 which seems like a reasonable value.
As the batch size has to be a divisor of the chunk size, the
latter was changed to 200 (formerly 250). To compensate  for the smaller
chunk size the initial maximum number of games for an sprt test is changed
to 2000 times the chunk size (formerly 1000 times the chunk size).

One unittest is updated.