Enable a new algorithm for choosing SPSA hyperparameters.
This PR adds a checkbox to the "New Test" form. Checking this option
will rewrite the hyperparameters furnished by the tuning branch in
such a way that the SPSA tune will finish within 0.5 Elo from the optimum
(with 95% probability) assuming the function 'parameters->Elo'
is quadratic and varies 2 Elo per parameter over each specified
parameter interval. The hyperparameters are relatively conservative
and their performance will degrade gracefully if the actual variation
is different.  The theoretical basis for choosing the hyperparameters is
given in this document

https://github.com/vdbergh/spsa_simul/blob/master/doc/theoretical_basis.pdf

The formulas can be checked by simulation which is done here

https://github.com/vdbergh/spsa_simul

It is important for the current version of the algorithm to be able to
estimate the draw ratio. The PR uses a heuristic method to do so which
makes various assumptions: book='noob_3moves.epd', contempt=24,
threads=1 and no nodestime.