Finding the default net from sources.
the changes needed to embed the net in the binary:
https://github.com/official-stockfish/Stockfish/pull/3070

need a change in the magic for finding the default net from sources on both the server and worker. This should be merged before we can merge the above PR

For backwards compatibility we now need to look for the name of the default net in two places.