Fix SyntaxWarning: "is" with a literal
https://docs.python.org/3.8/whatsnew/3.8.html#changes-in-python-behavior

The compiler now produces a SyntaxWarning when identity checks (is and is not)
are used with certain types of literals (e.g. strings, numbers).
These can often work by accident in CPython, but are not guaranteed by the
language spec.
The warning advises users to use equality tests (== and !=) instead.