More stats auditing.
Display the batch size on the raw statistics page.
Add the LLR jumps to the raw statistics page.

Sometimes a test finishes with an LLR of >3 (without it having been
purged). We can now check if this is reasonable.

While at it, also fix an unrelated issue with run['args']
having a different structure in very old tests.