List SPRT bounds in "normalized Elo" on the raw stats page.
Normalized Elo is inversely proportional
to the square root of the number of games
it takes on average to detect a given strength
difference with a given level of significance.

In case of a perfectly balanced book and
small Elo differences normalized Elo is
approximately equal to

logistic_elo/sqrt(1-draw_ratio)

In other words if the draw ratio is zero,
the book is perfectly balanced
and Elo differences are small then
normalized Elo and logistic Elo coincide.

The theory behind normalized Elo is discussed here

http://hardy.uhasselt.be/Fishtest/normalized_elo.pdf

Compared to loc. cit. we have settled on using
an additional normalization factor

800/log(10)=347.43558552260146

which has the effect that normalized Elo and logistic Elo
are easier to compare (as explained above).

We also convert the former sensitivity field to normalized Elo.

Hopefully using this conversion the field will become
easier to interpret.

The purpose of the sensitivity field is to compare opening books.
The opening book influences both Elo and the draw ratio. By expressing
the result of a test in normalized Elo one can compare different
books.