Use commit sha to retrieve net.
When resheduling a test, fishtest previously used the cached commit sha to retrieve
the code to execute, but fetched the net using only the branch name.

This leads to race conditions if the default net of the corresponding branch was changed.
To circumvent the issue use the commit sha also to fetch the net.