Decrease base nps
The recently added new net arch introduced a slowdown. On my system it is 6.8% slower. In the commit it is said to be 15% to 20% slower. I guess this could vary quite a lot between different CPU instruction sets. For now I decreased base_nps by 10% but there are arguments to be made to change it less/more depending if we change on most commonly supported instruction set on fishtest, slowest arch or fastest arch etc.
I hope @vondele and @snicolet have recommendations on how to deal with this matter.

Update views.py

Update delta_update_users.py