Correct the chi2 p-value computation for an unbalanced book.
Currently the chi2 test assumes that the games in a task are independently
drawn from a trinomial distribution. This assumption is no longer
valid for an unbalanced book. This is very visible from the p-values
which are often > 99%. Under the null hypothesis p-values should be
uniformly distributed.

Since currently the ww and ll frequencies will be too small for
the chi2 test to be valid, we coalesce them with the wd and ld frequencies.

The main test page still shows the trinomial wdl frequencies. This will be
addressed in a separate PR.