Enable fake net upload in a nice way.
This is to make the installation of a local fishtest (for testing)
a bit easier by avoiding the need of having a local cdn as well.

If the fishtest user has a file 'fake_net_upload' in its home
directory then an "uploaded" net will be recorded in the local database
but it will not be uploaded to any cdn.

It is the assumption that in that case the net already exists
in the official stockfish cdn.