Generate tasks dynamically.
We start with an empty task list and create new tasks as needed.

Detailed changes:

- Refactor "update_task".

The main change is that we do not inspect the task list to see
if a test is finished.

- Bug fix

Make task["active"]=False after a broken update.
I think this fixes a bug in the current Fishtest code.

- Try to create 60 min tasks using "estimate_game_duration",
"concurrency" and "threads".

The 60 min is configurable by the variable self.task_duration in
rundb.py. Currently its value is 3600.

- Make sure that the task_size is not zero for very long TC.

- Filter workers participating in the chi2 test.

We filter out the workers that have cells with expected
frequencies <=5. This is the recommended practice.

With the current UHO book this means that at LTC
a worker should contribute around 50 games to make the
cut.

We do not display residuals for filtered workers.