Sort "task_runs" in request_update() at the correct place.
Bugfix: the current code (like the old code) may not pick
up changes of run["cores"] due to finished tasks.

More precisely: if a cached version of task_runs is not sorted
at the beginning of request_task() its first entry may be a run
which does not have the lowest value of run["cores"]/run["itp"]
because it may be that since the last call to request_task()
one or more tasks in other runs finished (handled in update_task()).

Also some cosmetic changes around the request_task()/update_task()
code.