A worker fingerprint that should be mostly unique accross restarts
This should make it possible to more easily track workers that
are frequently restarted.

The fingerprint is shown on the machines table instead of the UUID
(but it contains the former UUID field).

The fingerprint consists of 3 fixed parts which are respectively
hashed versions of:

- The output of uuid.getnode() (normally the MAC address).
- The pathname of the worker script.
- The config options (even when modified by the command line).

The fourth part is the first field of the UUID (which is itself
randomly generated).