Try to find a new run for a worker in request_task().
Currenly there is the issue that high core workers always get
the same run while their residual keeps steadily rising.

This is not good. We want a good hardware mix for every test.

This PR moves the last run seen by the worker to the end of the queue.

Also: make sure not to return a task from a finished SPRT run.

Not live tested.