Get rid of worker_key altogether.
This should be a non functional conceptual simplification.

The worker key tries to solve two unrelated problems:

(1) Long ago workers did not have to a unique key, so a substitute
has to be provided for the chi2 test.

(2) The unique key is not suitable for displaying, and perhaps should
be kept secret as well. So the worker key is displayed instead.

Now we simply provide two functions

unique_key(worker_info)
worker_name(worker_info)

corresponding to (1,2).