Document and refactor request_task.
In master "request_task()" is a large for loop which contains
a large if. We get rid of the if. This makes the code
much more readable.

The current structure is

run_found=False
for run in self.task_runs:
     # comment1
     if excluding_condition1:
         continue
     # comment2
     if excluding_condition2:
         continue
     ...
     run_found=True
     break

if not run_found:
     return

(now create a new task for the run that was found)

This should be a non-functional change.