Refactor purging logic and overhaul the residuals computation.
The computation of bad_workers is now done inside purge_run(),
where it logically belongs.

As a side effect of the refactoring it is now easy to have
different purging criteria for auto purging and manual purging.

As a proof of concept we have set the purging criteria for
manual purging to p=0.01 and residual=4.5, whereas for auto purging
(disabled by default) we have set them to p=0.001 and residual=7.0.

We also use a different logic for computing residuals because the
previous residuals (defined as the maximum of the absolute values
of the Pearson adjusted residuals in a row) have an unknown distribution.
Now we compute (conceptually) for every row in the chi2 table the
chi2 value of the 2x3 table obtained by that row and the sum of
the other rows. This chi2 value follows a chi2 distribution with
two degrees of freedom (under the null hypothesis that the workers are
homogeneous). To keep some compatibility with the previous
residuals, we convert the chi2 value to a standard normal
value.

To give an indication of the severity of a residual, Fishtest
uses a coloring scheme. However this coloring scheme was created
when far fewer workers were active. These days most of the
yellow and red residuals are just the result of statistical
fluctuations and as a result nobody pays any attention to them.

In this PR we update the coloring scheme. More precisely we color
residuals according to the p=0.05 and p=0.01 thresholds, but
using the Bonferroni correction, which divides these p-values
by the number of workers.