Bug fix for purges with crashes and time losses.
The new purging logic purges workers (based on their residuals), not tasks.
So the current code which sets the residual to 8 for a task with several
crashes has no effect anymore (since it does not change the residual of the worker).

The implemented solution is to start purge_run() with a loop which filters
tasks with crashes and time losses, and then to compute the worker
based residuals (since these will be affected by the preliminary filtering).

Fixes #1058.