Various cleanups and refactorings.
FLEET is renamed "fleet". It turns out that it does not need
to be global.

ALIVE, RUN and TASK_ID are packed in a dictionary "current_state"
which is passed as a parameter to the subroutines that need it.

Extra comments in main() documenting the various initialization
stages.

Remove some useless delays from worker.py:worker(). There is already
a delay in the main loop. As a bonus worker() no longer needs to
know about the fleet variable.

Finish the main loop cleanly without invoking os._exit().

Make Ctrl-C work immediately in all cases.

Rename worker() as fetch_and_handle_task().