Give server feedback for failed tasks.
Post a message in the server log.

Post a message in the actiondb.

Introduce "WorkerException" for standard exceptions.

One does not want record all exceptions on the server since
this would include things like syntax errors, potentially even
leading to privacy issues.

If one raises WorkerException(reason) in run_games() then the reason will
be recorded on the server using the "failed_task" api. Exceptions caused
by syntax errors etc... are of course not of the type WorkerException.
Such exceptions will be recorded on the server with the file name and
the line number where the exception occurred.

Also: fix a bug in pgn uploading created by the last commit.