Stop worker if another is running in the folder
We called setup_engine() because the engine was not cached.
Only another worker running in the same directory can have built the engine.
There is still a little probability of race condition
between the existence check and the move of the file.

A proper solution to the problem of multiple workers running in the same
directory requires a check at the start of the worker (eg using a file lock).

While here build the source code path with the robust way used in updater.