Do not mask Ctrl-C when wrapping a call.
With this PR the correct message is used even if the user
presses Ctrl-C (in which case "Terminated by signal" is used).
This is done by equipping WorkerException with an optional
argument e. If e is of the type WorkerException or a subclass
of it then it will be forwarded. We use some __new__ hackery for this.

This PR establishes a pattern for cases where forwarding inner
exceptions is useful.

We use it at three places

- in requests_get()
- in requests_post()
- while wrapping the main make command.

Also for consistency: two more FatalException when handling cutechess.