A very primitive locking implementation.
We don't want two copies of the worker running in the
same directory.

The most natural way to achieve this would be to use a lock file.

Unfortunately Python doesn't have a cross platform file locking api.
So we just look for the existence of a lock file.
We do this when starting the worker (which is racy) and also
(as a fallback) before requesting a new task
(in that case we also check that the content of the lock file
is different from the current pid).