Make lock file system more robust.
We may have several processes reading from and writing to the lock file
at the same time with potentially unpredictable consequences
(especially on windows). We now refactor the file IO operations
and put them in a try: except: bracket.

We also change the logic:

At the start of the worker we tolerate an invalid or non-existing
lock file since we intend to replace it with one of our own.
Once  we have started, we now only accept a valid lock file
containing our own PID.

Also: the name of the lock file is not a global variable anymore.

Also: send stderr to devnull on Linux. Like it is done on
Windows.