Do more validation of api arguments in api.py.
This allows for simplication of downstream code. Notably
"failed_task()", "update_task()" and "request_spsa()".

We do the validation when parsing the arguments. For example
self.task_id() will throw a HTTPBadRequest() exception if it
does not succeed in returning a task_id corresponding to an
existing task. So downstream code may assume that a task_id
argument is valid.

The same mechanism has allowed us to simplify api_stop_run
which had recently ballooned because of argument validation.

To check that the api arguments are syntactically correct we
provide a rudimentary, but generic, "schema" implementation in
utils.py:validate().