Do not make dead tasks visible by default in the events log.
This is similar to system events.

A dead task may for example be caused by a user closing their
laptop. Or a colab worker being disconnected.

Dead tasks do not carry a lot of information (they have no
associated message) and there are a lot of them.

But the fact that a task was declared dead can explain why a pgn file
was not uploaded. So it might still be useful to have a record of
them.