Imply use_all_cores=False if lowering concurrency.
Currently, if you launch the client once with use_all_cores, it
will save this in the configuration. If you then relaunch it with
a lower --concurrency, that will get ignored because the saved
use_all_cores=True overrules it.

The cleanest fix (that is still backwards compatible and avoids
introducing new flags) appears to be to make a concurrency that's
less than maximum imply use_all_cores=False.

Fixes #1184.