Some tweaks around option handling.
1) We try to fix

https://github.com/glinscott/fishtest/issues/1184

The semantics of --use_all_cores is now as follows:

--use_all_cores True

(1) allows the use of all cores;
(2) makes the default concurrency equal to the number of cores.

The default can be overridden by supplying an explicit
--concurrency option.

--use_all_cores False disallows the use of all cores.

2) We validate config file entries and command line parameters.

In particular we make sure that in the option object returned by
setup_parameters() all attributes have the expected type. This
makes the downstream code simpler.

3) We switch from optparse to argparse. optparse is deprecated.