Some tweaks around option handling.
1) This PR started by trying to fix.

https://github.com/glinscott/fishtest/issues/1184

It turned that there are fundamental issues with the
semantics of --use_all_cores. So it was decided
to deprecate --use_all_cores and to introduce
--concurrency max as a clearly defined substitute for
--use_all_cores True.

Some backward compatibility is temporarily retained, but
--use_all_cores will be removed in the future to simplify
the worker code.

2) We now strictly validate config file entries and command
line parameters.

In particular we make sure that in the options object returned by
setup_parameters() all attributes have the expected type. This
makes the downstream code simpler.

3) We switch from optparse to argparse. optparse is deprecated.

4) We now check, at the startup of the worker, that the
password, supplied via the config file or on the command line,
is correct. If it is not then we present a password dialog.

This validation step is skipped when the flag --no_validation is
present on the command line. In combination with --only_config
this is useful when there is no network access.

5) --only_config is now a flag. In contrast --fleet remains
a boolean since all three possibilities (True/False or omitted)
are meaningful.