Do not delete a task after purging but mark it as bad.
With the current code a purged task is deleted from the task
list. This means that the numbering of tasks changes.

One consequence of this is that links to tasks in the event log
may become invalid.

Another issue is that task_id's held by the worker become invalid.
This is unlikely to cause problems but it is still not nice
and could lead to bugs later.

In this PR we mark a purged task as bad. For safety we also make
it non-active and set all its stats to zero.