Introduce generic logging function with timestamp.
For api logging we now print something user friendly on the
terminal and for logging we use something more machine parsable.

We also log run/task_id since there may be a correlation between
timings and the length of the run.

Also: print current time before fetching a task for people that
run the worker under nohup.