Create "failed", "deleted", "is_green", "is_yellow" flags in new_run.
Queries in mongodb for missing flags are a bit messy.
This is the first step at fixing this. The next step is to
loop through the runs collection and set/unset these flags as needed.

Also: we document where the flags

"finished", "failed", "deleted", "is_green", "is_yellow" are set
and unset.