Make run_table.mak more parametrizable.
If we want to implement a general query system
then we need a way to present the results.
run_table.mak almost fits the bill
but it is not very self contained. In this
PR we fix this.

As an example we make all sections on the first page
collapsible. This is now trivial to do.

Also:

Delete obsolete copy of jquery.cookie.js and
get it from a cdn instead. Added the
appropriate SRI hash.

Use slideToggle also for the machines table.

Remove some inline svg and create images instead.

Workaround for a bug in pyramid.request.cookies.
The bug happens with Chrome if there are cookies with the same
name and different paths. Then the server and the browser
disagree on which cookie to use. Currently this does not happen
but I kept the code around since the issue could strike
again in the future.